Move `GateTarget::from_target_str` to `GateTarget.h`
